requirements.txt
async-timeout==4.0.3
azure-appconfiguration==1.5.0
azure-core==1.29.7
azure-functions==1.18.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
isodate==0.6.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
redis==5.0.1
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
